+++ src/core/kernel/opencl/optimizer.hh
@@ -9,7 +9,7 @@ namespace opencl_kernel
 inline std::string optimizer_kernel_code{R"(
 
 // #include "../../core/kernel/opencl/typedef.hh"
-typedef double value_type;
+typedef float value_type;
 
 kernel void naive(
     value_type alpha,
