+++ src/core/kernel/opencl/tanh.hh
@@ -9,7 +9,7 @@ namespace opencl_kernel
 inline std::string tanh_kernel_code{R"(
 
 // #include "typedef.hh"
-typedef double value_type;
+typedef float value_type;
 
 kernel void forward(
     global value_type const* in,
